:goal_net: feat: refactor and add error tracking
